move mesh offset info for easier forward declare
