+++ kimera_pgmo/CMakeLists.txt
@@ -49,6 +49,7 @@ add_library(
   src/mesh_deformation.cpp
   src/mesh_delta.cpp
   src/mesh_frontend_interface.cpp
+  src/mesh_offset_info.cpp
   src/mesh_traits.cpp
   src/kimera_pgmo_interface.cpp
   src/sparse_keyframe.cpp)
