+++ kimera_pgmo/src/mesh_delta.cpp
@@ -12,19 +12,6 @@ namespace kimera_pgmo {
 
 using TrackingInfo = MeshDelta::TrackingInfo;
 
-size_t MeshOffsetInfo::toGlobal(size_t local_idx) const {
-  return local_idx + archived_vertices;
-}
-
-size_t MeshOffsetInfo::toLocal(size_t global_idx) const {
-  return global_idx - archived_vertices;
-}
-
-void MeshOffsetInfo::RemapInfo::addIndex(size_t idx) {
-  min_index = std::min(min_index, idx);
-  max_index = std::max(max_index, idx);
-}
-
 MeshDelta::MeshDelta(const TrackingInfo& info) : info(info) {}
 
 size_t MeshDelta::addVertex(const traits::Pos& pos,
